pht: account for nextmig usefully in iteration
This avoids scanning migrated hashchains in both table_next() and
pht_nextval(), and skips right past to nextmig when landing within the
hashchain under migration, which is optimal.

Also, this is the kind of shit what can be done when tests cover things like
an off-by-one in pht_add() that might otherwise be introduced unnoticed.