pht: support small tables
This is mostly a cleanup changeset in anticipation of the common-bits
mechanism, up next.

To wit, since two bits are exposed from the common mask for the first incoming
item of an empty hash table, additional items are nigh guaranteed to require
exposure of further bits and therefore reallocation of the table. This means
that having a MIN_TABLE size above 1 only increases the amount of migration
scanning while providing no performance advantage, so the revised new_table()
starts at 1 and uses powers of two thereafter.