pht: add rehash-avoiding fast migration
When migrating a perfect item to a table of same or smaller size, the position
of that perfect item in the source corresponds to its position in the
destination (shifted down), so it can be inserted without rehashing. When the
destination table is instead larger, a perfect item may regardless migrate
without rehashing if its possible destination range[0] doesn't already contain
a hash-chain break, and it gives up its perfect bit.

For the time being, imperfect items are always migrated by rehashing.
Consequently, perfect items are made imperfect by a first migration to a
larger table, and may then be re-perfected the second time around.

This changeset adds fast_migrate() which is fast in the sense of not touching
likely-cold cache lines, but slow in terms of cycles spent under favourable
caching.

[0] see immediately preceding cset