pht: grow hash values at the bottom
With this change, S_i maps to T_i and T_(i+1) when #T = #S * 2, where
previously it'd map to T_i and T_(i+#S) which is less cache-friendly.

As this means using the higher bits of the raw hash value which are
substantially less mixed for many hash functions, we also compensate by
xor'ing the hash value with itself rotated to the right some arbitrary number
of bits.

A similar mixing step in stash_bits() was changed so that its result no longer
depends on table size. This makes stash bits transferrable between tables of
different sizes.