bench: add no-fork mode
no-fork mode runs benchmarks in the main process and uses the child process to
gather and report results. This causes e.g. asserts hit by the benchmark to
drop a proper corefile, instead of one from the parent. Benchmark results in
no-fork mode are invalid for not having a level VM playing field between
variants.

Also allow specification of the words file on the command line.