bench: increase tally depth
It was previously giving very unflattering numbers for CCAN htable due to its
results' wide range. This is better, but perhaps the error values should be
baked into the presented results rather than leave it to the reader.

Certainly CCAN htable doesn't take 580 clocks per back-to-back add, median.
That's absurd.